moving BuildTestError to buildtest/utils/exceptions.py Refactor import statements Add regtest for BuildTestError
