+++ buildtest/utils/command.py
@@ -1,26 +1,5 @@
 import subprocess
 
-
-class BuildTestError(Exception):
-    """Class responsible for error handling in buildtest. This is a sub-class
-    of Exception class."""
-
-    def __init__(self, msg, *args):
-        """Constructor Method.
-
-        :param msg: message to print
-        :type msg: str, required
-        :param args:
-        :type args:
-        """
-        if args:
-            msg = msg % args
-        self.msg = msg
-
-    def __str__(self):
-        return repr(self.msg)
-
-
 class BuildTestCommand:
     """Class method to invoke shell commands and retrieve output and error. This class
     makes use of **subprocess.Popen()** to run the shell command. This class has no
